requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-aws>=0.2.0

mcp>=0.9.0
httpx>=0.24.0

weaviate-client>=4.0.0
pandas>=2.0.0
numpy>=1.24.0

google-generativeai>=0.8.0
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-storage>=2.22.0
db-dtypes>=1.1.1
pyarrow>=14.0.0
boto3>=1.34.0

pyyaml>=6.0
python-dotenv>=1.0.0

pytest>=8.0.0
pytest-asyncio>=0.23.0

pydantic>=2.0.0
typing-extensions>=4.0.0
